big update

+++ users/views.py
@@ -29,7 +29,7 @@ class UserLoginView(views.APIView):
         if user and check_password(password, user.password):
             refresh_token = RefreshToken.for_user(user=user)
             access_token = refresh_token.access_token
-            return Response({'token': str(access_token)}, status=status.HTTP_200_OK)
+            return Response({'token': str(access_token), 'user_id': user.id, 'user_role': user.role}, status=status.HTTP_200_OK)
         return Response({'message': ''}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -45,13 +45,15 @@ class UserRegistrationView(views.APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-# @authentication_classes([UserAuthentication])
-# @permission_classes([IsAuthenticated])
+@authentication_classes([UserAuthentication])
+@permission_classes([IsAuthenticated])
 @api_view(['GET'])
-@permission_classes([AllowAny])
 def get_users(request):
-    # user = get_user_from_token(request.headers)
-    # if user.role == User.ADMIN:
+    user = get_user_from_token(request.headers)
+    if user.role != User.ADMIN:
+        return Response({"error": "You are not authorized to perform this action."},
+                        status=status.HTTP_403_FORBIDDEN)
+    
     users = User.objects.all()
     all_users = [
         {
@@ -60,16 +62,15 @@ def get_users(request):
             'first_name': user.first_name,
             'last_name': user.last_name,
             'email': user.email,
+            'avatar': user.avatar.url if user.avatar else None,
             'role': user.role,
             'created_at': user.created_at,
             'updated_at': user.updated_at,
         }
         for user in users
     ]
+    
     return Response(all_users, status=status.HTTP_200_OK)
-    # serializer = UserSerializer(users, many=True)
-    # return Response(serializer.data, status=status.HTTP_200_OK)
-    # return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['POST'])
@@ -86,11 +87,15 @@ def logout_user(request):
     except Exception as e:
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-
+@authentication_classes([UserAuthentication])
+@permission_classes([IsAuthenticated])
 @api_view(['GET'])
-@permission_classes([AllowAny])
 def get_users_by_role(request, role):
     try:
+        user = get_user_from_token(request.headers)
+        if user.role != User.ADMIN:
+            return Response({"error": "You are not authorized to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         users = User.objects.filter(role=role)
         # Manually select the desired fields
         users_data = [
@@ -110,13 +115,18 @@ def get_users_by_role(request, role):
     except Exception as e:
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-
-@api_view(['PUT'])
+@authentication_classes([UserAuthentication])
 @permission_classes([IsAuthenticated])
+@api_view(['PUT'])
 def update_user(request, user_id):
     try:
         user = User.objects.get(pk=user_id)
-        serializer = UserSerializer(instance=user, data=request.data, partial=True)
+        
+        # Extract only the allowed fields from request.data
+        allowed_fields = ['first_name', 'last_name', 'email', 'role', 'date_of_birth', 'avatar', 'description', 'mobile', 'social_media_url', 'location']
+        data_to_update = {field: request.data.get(field) for field in allowed_fields if field in request.data}
+        
+        serializer = UserSerializer(instance=user, data=data_to_update, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
@@ -127,15 +137,15 @@ def update_user(request, user_id):
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
+@authentication_classes([UserAuthentication])
+@permission_classes([IsAuthenticated])
 @api_view(['DELETE'])
-# @permission_classes([IsAuthenticated])
-@permission_classes([AllowAny])
 def delete_user(request, user_id):
     try:
-        # user = get_user_from_token(request.headers)
-        # if user.role != User.ADMIN:
-        #     return Response({"error": "You are not authorized to perform this action."},
-        #                     status=status.HTTP_403_FORBIDDEN)
+        user = get_user_from_token(request.headers)
+        if user.role != User.ADMIN:
+            return Response({"error": "You are not authorized to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
 
         user = User.objects.get(pk=user_id)
         user.delete()
@@ -146,10 +156,15 @@ def delete_user(request, user_id):
         return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
+@authentication_classes([UserAuthentication])
+@permission_classes([IsAuthenticated])
 @api_view(['GET'])
-@permission_classes([AllowAny])
 def get_user_by_id(request, user_id):
     try:
+        user = get_user_from_token(request.headers)
+        if user.role != User.ADMIN:
+            return Response({"error": "You are not authorized to perform this action."},
+                            status=status.HTTP_403_FORBIDDEN)
         user = User.objects.get(pk=user_id)
         user_data = {
             'id': user.id,
@@ -158,6 +173,12 @@ def get_user_by_id(request, user_id):
             'last_name': user.last_name,
             'email': user.email,
             'role': user.role,
+            'date_of_birth': user.date_of_birth,
+            'avatar': user.avatar.url,
+            'description': user.description,
+            'mobile': user.mobile,
+            'social_media_url': user.social_media_url,
+            'location': user.location,
             'created_at': user.created_at,
             'updated_at': user.updated_at,
         }
